add few handlers

+++ internal/handler/offering.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"fmt"
 	"net/http"
 	"terminer/internal/models"
 
@@ -55,13 +56,38 @@ func (h *Handler) DeleteService(c *gin.Context) {
 		return
 	}
 
-	if err := h.services.Offering.DeleteService(input.UUID); err != nil {
+	userId, err := getUserId(c)
+	if err != nil {
 		NewErrorResponse(c, http.StatusInternalServerError, err.Error())
 		return
 	}
-	c.JSON(http.StatusOK, map[string]interface{}{
-		"message": "ok",
-	})
+	isAdmin, err := h.services.User.IsAdmin(userId)
+	if err != nil {
+		NewErrorResponse(c, http.StatusInternalServerError, err.Error())
+		return
+	}
+	owner, err := h.services.GetServiceOwner(input.UUID)
+	fmt.Println("from handler", owner)
+	if isAdmin == true || owner == userId {
+		if err := h.services.Offering.DeleteService(input.UUID); err != nil {
+			NewErrorResponse(c, http.StatusInternalServerError, err.Error())
+			return
+		}
+
+		c.JSON(http.StatusOK, map[string]interface{}{
+			"message": "ok",
+			"status":  "deleted",
+		})
+
+	} else {
+
+		c.JSON(http.StatusOK, map[string]interface{}{
+			"message": "User not owner or admin",
+			"status":  "Not deleted",
+		})
+
+	}
+
 }
 
 func (h *Handler) GetTypes(c *gin.Context) {
@@ -72,5 +98,39 @@ func (h *Handler) GetTypes(c *gin.Context) {
 		return
 	}
 	c.JSON(http.StatusOK, types)
+
+}
+
+func (h *Handler) CreateServiceType(c *gin.Context) {
+	var input models.ServiceType
+	if err := c.BindJSON(&input); err != nil {
+		NewErrorResponse(c, http.StatusBadRequest, err.Error())
+		return
+	}
+	if err := h.services.Offering.CreateServiceType(input); err != nil {
+		NewErrorResponse(c, http.StatusInternalServerError, err.Error())
+		return
+	}
+	c.JSON(http.StatusOK, map[string]interface{}{
+		"message": "ok",
+		"status":  "created",
+	})
+}
+
+func (h *Handler) GetMyServices(c *gin.Context) {
+
+	userId, err := getUserId(c)
+	if err != nil {
+		NewErrorResponse(c, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	services, err := h.services.Offering.GetMyServices(userId)
+	if err != nil {
+		NewErrorResponse(c, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	c.JSON(http.StatusOK, services)
 
 }
